Working transaction management
Transaction management set up using JTA's transaction management with
annotations such that either both insertions succeed or both fail, but
there is not a situation where one succeeds and the other fails.
Can be testing by entering a POST request to create a transaction with
text of < 10 and finding the insertion in both tables in the db, versus
an entry of > 10 whereby both are missing.